[FEAT] 로그 스택 붙임
