+++ logs/log.json
@@ -1,0 +1,3 @@
+{"@timestamp":"2023-01-13T17:48:08.971Z","log.level":"error","message":"[DELETE] /goal/1111111111 || Request Body: {} uid: 15 || error: Error: \nInvalid `prisma.goal.delete()` invocation in\n/Users/mk-mac-386/Documents/Cracker-Server/src/service/goalService.ts:77:40\n\n  74 };\n  75 \n  76 const deleteGoal = async (goalId: number) => {\n→ 77   const data = await prisma.goal.delete(\nAn operation failed because it depends on one or more records that were required but not found. Record to delete does not exist.","ecs":{"version":"1.6.0"}}
+{"@timestamp":"2023-01-13T17:48:31.644Z","log.level":"debug","message":"[DELETE] /goal/478 || Request Body: {} uid: 15","ecs":{"version":"1.6.0"}}
+{"@timestamp":"2023-01-13T17:53:37.447Z","log.level":"error","message":"[DELETE] /goal/478 || Request Body: {} uid: 15 || error: Error: \nInvalid `prisma.goal.delete()` invocation in\n/Users/mk-mac-386/Documents/Cracker-Server/src/service/goalService.ts:77:40\n\n  74 };\n  75 \n  76 const deleteGoal = async (goalId: number) => {\n→ 77   const data = await prisma.goal.delete(\nAn operation failed because it depends on one or more records that were required but not found. Record to delete does not exist.","ecs":{"version":"1.6.0"}}
