+++ package.json
@@ -21,14 +21,15 @@
     "@types/node-schedule": "^2.1.0",
     "@typescript-eslint/eslint-plugin": "^5.47.1",
     "@typescript-eslint/parser": "^5.47.1",
-    "nodemon": "^2.0.20",
-    "husky": "^8.0.0"
+    "husky": "^8.0.0",
+    "nodemon": "^2.0.20"
   },
   "dependencies": {
     "@elastic/ecs-winston-format": "^1.3.1",
     "@prisma/client": "^4.8.0",
     "@types/body-parser": "^1.19.2",
     "@types/supertest": "^2.0.12",
+    "@types/winston": "^2.4.4",
     "axios": "^1.2.2",
     "chai": "^4.3.7",
     "dayjs": "^1.11.7",
@@ -38,7 +39,7 @@
     "jsonwebtoken": "^9.0.0",
     "node-schedule": "^2.1.0",
     "prisma": "^4.8.0",
-    "winston": "^3.8.2",
-    "supertest": "^6.3.3"
+    "supertest": "^6.3.3",
+    "winston": "^3.8.2"
   }
 }
