+++ src/controller/goalController.ts
@@ -9,6 +9,7 @@ import date from "../modules/date"
 import boxCounter from "../modules/boxCounter";
 import achievedError from "../constants/achievedError";
 import { validationResult } from "express-validator";
+import { debugLog, errorLog } from "../logger/logger";
 
 const createGoal = async (req: Request, res: Response) => {
   const error = validationResult(req);
@@ -26,22 +27,34 @@ const createGoal = async (req: Request, res: Response) => {
     const startedAt = date.getNowPlus9()
 
     const data = await goalService.createGoal(userId, createGoalDTO, startedAt);
-
+    
+    debugLog(req.originalUrl, req.method, req.body, req.user?.userId);
     return res.status(sc.OK).send(success(sc.OK, rm.CREATE_GOAL_SUCCESS, data));
+  
   } catch (error) {
+  
     slack.sendErrorMessageToSlack(req.method.toUpperCase(), req.originalUrl, error, req.user?.userId);
+    errorLog(req.originalUrl, req.method, req.body, error, req.user?.userId);
     return res.status(sc.INTERNAL_SERVER_ERROR).send(fail(sc.INTERNAL_SERVER_ERROR, rm.INTERNAL_SERVER_ERROR)); 
+  
   }
 }
 
 const deleteGoal = async (req: Request, res: Response) => {
   try {
+  
     const { goalId } = req.params;
     const deletedGoalId = await goalService.deleteGoal(+goalId);
+    
+    debugLog(req.originalUrl, req.method, req.body, req.user?.userId);
     return res.status(sc.OK).send(success(sc.OK, rm.DELETE_GOAL_SUCCESS, { "goalId": deletedGoalId }));
+  
   } catch (error) {
+
     slack.sendErrorMessageToSlack(req.method.toUpperCase(), req.originalUrl, error, req.user?.userId);
+    errorLog(req.originalUrl, req.method, req.body, error, req.user?.userId);
     return res.status(sc.INTERNAL_SERVER_ERROR).send(fail(sc.INTERNAL_SERVER_ERROR, rm.INTERNAL_SERVER_ERROR)); 
+  
   }
 };
 
@@ -56,10 +69,17 @@ const updateGoal = async (req: Request, res: Response) => {
   const { goalId } = req.params;
 
   try {
+  
     const updatedGoalId = await goalService.updateGoal(+goalId, updateGoalDTO);
+
+    debugLog(req.originalUrl, req.method, req.body, req.user?.userId);
     return res.status(sc.OK).send(success(sc.OK, rm.UPDATE_GOAL_SUCCESS, { "goalId": updatedGoalId }));
+  
   } catch (error) {
+
     slack.sendErrorMessageToSlack(req.method.toUpperCase(), req.originalUrl, error, req.user?.userId);
+    errorLog(req.originalUrl, req.method, req.body, error, req.user?.userId);
+    
     return res.status(sc.INTERNAL_SERVER_ERROR).send(fail(sc.INTERNAL_SERVER_ERROR, rm.INTERNAL_SERVER_ERROR)); 
   }
 
@@ -76,9 +96,15 @@ const keepGoal = async (req: Request, res: Response) => {
   try {
     const keptAt = date.getNowPlus9()
     const keptGoalId = await goalService.keepGoal(+goalId, isOngoing, keptAt);
+  
+    debugLog(req.originalUrl, req.method, req.body, req.user?.userId);
     return res.status(sc.OK).send(success(sc.OK, rm.KEEP_GOAL_SUCCESS, { "goalId": keptGoalId }));
+  
   } catch (error) {
+  
     slack.sendErrorMessageToSlack(req.method.toUpperCase(), req.originalUrl, error, req.user?.userId);
+    errorLog(req.originalUrl, req.method, req.body, error, req.user?.userId);
+  
     return res.status(sc.INTERNAL_SERVER_ERROR).send(fail(sc.INTERNAL_SERVER_ERROR, rm.INTERNAL_SERVER_ERROR)); 
   }
 
@@ -111,10 +137,14 @@ const getHistoryByGoalId = async (req: Request, res: Response) => {
       "emptyBoxCount": boxCounter.getEmptyBoxCount(thisMonthCount)
     }
 
+    debugLog(req.originalUrl, req.method, req.body, req.user?.userId);
     return res.status(sc.OK).send(success(sc.OK, rm.GET_GOAL_SUCCESS_FOR_HISTORY, data));
   
   } catch (error) {
+    
     slack.sendErrorMessageToSlack(req.method.toUpperCase(), req.originalUrl, error, req.user?.userId);
+    errorLog(req.originalUrl, req.method, req.body, error, req.user?.userId);
+    
     return res.status(sc.INTERNAL_SERVER_ERROR).send(fail(sc.INTERNAL_SERVER_ERROR, rm.INTERNAL_SERVER_ERROR));
   }
 }
@@ -124,19 +154,18 @@ const achieveGoal = async (req: Request, res: Response) => {
   const { goalId } = req.params;
   const { isAchieved } = req.body;
 
-  console.log("isAchieved ", isAchieved);
   if (!userId || isAchieved === null || isAchieved === undefined) {
     return res.status(sc.BAD_REQUEST).send(fail(sc.BAD_REQUEST, rm.NULL_VALUE));
   }
 
   let data;
   try {
-
     data = await goalService.achieveGoal(+goalId, isAchieved as boolean);
-    console.log("Data ", data);
+    debugLog(req.originalUrl, req.method, req.body, req.user?.userId);
 
   } catch (error) {
     slack.sendErrorMessageToSlack(req.method.toUpperCase(), req.originalUrl, error, req.user?.userId);
+    errorLog(req.originalUrl, req.method, req.body, error, req.user?.userId);
     return res.status(sc.INTERNAL_SERVER_ERROR).send(fail(sc.INTERNAL_SERVER_ERROR, rm.INTERNAL_SERVER_ERROR)); // 서버 내부 에러
   }
