+++ src/controller/mypageController.ts
@@ -3,6 +3,7 @@ import { sc, rm } from "../constants";
 import { fail, success } from "../constants/response";
 import { mypageService } from "../service";
 import slack from "../modules/slack";
+import { debugLog, errorLog } from "../logger/logger";
 
 const sortType = {
   ALL: "all",
@@ -24,10 +25,17 @@ const getMypageByUserId = async (req: Request, res: Response) => {
   }
 
   try {
+  
     const foundGoals = await mypageService.getGoalsForMypage(+userId, sort as string);
+    
+    debugLog(req.originalUrl, req.method, req.body, req.user?.userId);
     return res.status(sc.OK).send(success(sc.OK, rm.GET_GOALS_SUCCESS_FOR_MYPAGE, { "goals": foundGoals, "goalCount": foundGoals.length }));
+  
   } catch (error) {
+    
     slack.sendErrorMessageToSlack(req.method.toUpperCase(), req.originalUrl, error, req.user?.userId);
+    errorLog(req.originalUrl, req.method, req.body, error, req.user?.userId);
+
     return res.status(sc.INTERNAL_SERVER_ERROR).send(fail(sc.INTERNAL_SERVER_ERROR, rm.INTERNAL_SERVER_ERROR));
   }
